+++ cst438_register/src/main/java/com/cst438/controller/ScheduleController.java
@@ -156,22 +156,5 @@ public class ScheduleController {
 		return courseDTO;
 	}
 	
-	/*
-	 *  add a student to the system.With the input of the student email and name.
-	 */
-	@PostMapping("/student")
-	@Transactional
-	public Student addStudent(@RequestBody StudentDTO studentDTO) {
-		Student emailIdCheck = studentRepository.findByEmail(studentDTO.email);
-		if(studentDTO != null && emailIdCheck == null){
-			Student student = new Student();		
-			student.setName(studentDTO.name);
-			student.setEmail(studentDTO.email);
-			Student saveStudent = studentRepository.save(student);
-			return saveStudent;
-		}else{
-			throw  new ResponseStatusException( HttpStatus.BAD_REQUEST, "emailID is already exists ");
-		}
-		
-	}
+
 }
