+++ cst438_register/src/main/java/com/cst438/domain/StudentDTO.java
@@ -1,0 +1,7 @@
+package com.cst438.domain;
+
+public class StudentDTO {
+	public String name;
+	public String email;
+
+}
